\name{basal_area}
\alias{basal_area}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Basal Area Projection Equation
}
\description{
Preojects basal area for thinned or unthinned stands from an initial measurement or simulation period.
}
\usage{
basal_area(stand.df)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{stand.df}{
A data frame with the required stand values. See ??standsim for details.
}
}
\value{
Returns a value of basal area per hectare (m2)
}
\references{
Allen, M.G, C. Anton-Fernandez, and Rasmus Austrup. 2020. A Stand-Level Growth and Yeild model for Managed Norway Spruce forest in Norway. Scandianvian Journal of Forest Research (in review).
}
\author{
Micky G Allen II
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
