\name{dom_height}
\alias{dom_height}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Dominant Height Projection Equation
}
\description{
Projects stand dominant height from an initial measurement or simulation period.
}
\usage{
dom_height(stand.df)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{stand.df}{
A data frame with the required stand values. See ??standsim for details.
}
}

\value{
Returns a value of future stand dominant height in meters
}
\references{
Allen, M.G, C. Anton-Fernandez, and Rasmus Austrup. 2020. A Stand-Level Growth and Yeild model for Managed Norway Spruce forest in Norway. Scandianvian Journal of Forest Research (in review).
}
\author{
Micky G Allen II
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
