\name{perc_saw_volume_thin}
\alias{perc_saw_volume_thin}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Percentage of Volume in Sawtimber of Thinned Stands
}
\description{
Predicts the percentage of volume in the sawtimber product class in thinned stands.
}
\usage{
perc_saw_volume_thin(stand.df)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{stand.df}{
A data frame with the required stand values. See ??standsim for details.
}
}
\references{
Allen, M.G., C. Anton-Fernandez, and R. Astrup. 2020. Economical Impacts of Thinning in Managed Norway Spruce Forests in Norway. Scandinavian Journal of Forest Research (in review).
}
\author{
Micky G Allen II
}


