\name{perc_waste_volume}
\alias{perc_waste_volume}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Percentage of Volume in Waste Wood
}
\description{
Predicts the percentage of total stand volume in the waste product class.
}
\usage{
perc_waste_volume(stand.df)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{stand.df}{
A data frame with the required stand values. See ??standsim for details.
}
}
\references{
Allen, M.G., C. Anton-Fernandez, and R. Astrup. 2020. Economical Impacts of Thinning in Managed Norway Spruce Forests in Norway. Scandinavian Journal of Forest Research (in review).
}
\author{
Micky G Allen II
}

