\name{qmd}
\alias{qmd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Quadratic Mean Diameter Function
}
\description{
Calculates quadratic mean diameter (cm) from stand basal area (m2/ha) and the number of tree per hectare.
}
\usage{
qmd(stand.df)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{stand.df}{
A data frame with the required stand values. See ??standsim for details.
}
}
\author{
Micky G Allen II
}

