\name{survival}
\alias{survival}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Survival Projection Equation
}
\description{
Projects the number of trees per hectare from an initial measurements or simulation period.
}
\usage{
survival(stand.df)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{stand.df}{
A data frame with the required stand values. See ??standsim for details.
}
}
\value{
Returns a value of the future number of trees per hectare.
}
\references{
Allen, M.G, C. Anton-Fernandez, and Rasmus Austrup. 2020. A Stand-Level Growth and Yeild model for Managed Norway Spruce forest in Norway. Scandianvian Journal of Forest Research (in review).
}
\author{
Micky G Allen II
}


