\name{tree_reduction}
\alias{tree_reduction}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Reduction in Trees per Hectare following thinning
}
\description{
Predicts the residual number of trees per hectare after thinning.
}
\usage{
tree_reduction(stand.df)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{stand.df}{
A data frame with the required stand values. See ??standsim for details.
}
}
\value{
Returns a value of the residual number of trees following thinning.
%% ...
}
\references{
Allen, M.G, C. Anton-Fernandez, and Rasmus Austrup. 2020. A Stand-Level Growth and Yeild model for Managed Norway Spruce forest in Norway. Scandianvian Journal of Forest Research (in review).
}
\author{
Micky G Allen II
}

