\name{volume}
\alias{volume}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Stand level Volume Equation
}
\description{
Predicts the stand volume per hectare from other stand values.
}
\usage{
volume(stand.df)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{stand.df}{
A data frame with the required stand values. See ??standsim for details.
}
}
\value{
Returns a value of stand level volume per hectare (m3).
}
\references{
Allen, M.G, C. Anton-Fernandez, and Rasmus Austrup. 2020. A Stand-Level Growth and Yeild model for Managed Norway Spruce forest in Norway. Scandianvian Journal of Forest Research (in review).
}
\author{
Micky G Allen II
}

