\name{wholestandDF}
\alias{wholestandDF}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Convert LIST from standsim to a dataframe
}
\description{
Converts the LIST object generated from the function "standsim" into a dataframe.
}
\usage{
wholestandDF(fl)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fl}{
LIST generated from wholestand call.
}
}
\value{
Returns a dataframe
}

\author{
Micky G Allen II
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.


}

